# docs/swagger.yaml
basePath: /v1
definitions:
  Exchange_internal_model.ErrorResponseStruct:
    properties:
      code:
        type: integer
      error:
        description: پیام فنی (developer-friendly)
        type: string
      error_code:
        description: کد خطای اختصاصی (مثلاً validation_error)
        type: string
      message:
        description: پیام کاربر (user-friendly)
        type: string
      success:
        type: boolean
    type: object
...
  Exchange_internal_model.Response:
    properties:
      code:
        description: HTTP status code
        type: integer
      data:
        description: داده (در صورت موفقیت)
      error_code:
        description: کد یکتا برای خطا (اختیاری)
        type: string
      message:
        description: پیام (موفق یا خطا)
        type: string
      success:
        description: وضعیت موفقیت
        type: boolean
    type: object
...
  Exchange_internal_model.WalletDepositRequest:
    properties:
      amount:
        type: number
      currency_id:
        type: string
      user_id:
        type: string
    required:
    - amount
    - currency_id
    - user_id
    type: object
  Exchange_internal_model.WalletWithdrawRequest:
    properties:
      amount:
        type: number
      user_id:
        type: string
    required:
    - amount
    - user_id
    type: object
  internal_delivery_handler.VerifyRequest:
    properties:
      request_id:
        type: string
      user_input:
        type: string
    required:
    - request_id
    - user_input
    type: object
host: localhost:7000
info:
  contact: {}
  description: Exchange cryptocurrency trading platform
  title: Exchange API
  version: "1.0"
paths:
...
  /captcha:
    get:
      description: Generates a captcha image and returns it along with a unique request_id.
      produces:
      - application/json
      responses:
        "200":
          description: JSON response with request_id and base64 captcha string.
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "500":
          description: Failed to generate or store captcha
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      summary: Generate captcha image
      tags:
      - captcha
  /captcha/verify:
    post:
      consumes:
      - application/json
      description: Verify user input against the stored captcha text.
      parameters:
      - description: Captcha verification info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_handler.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Captcha is valid
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: Invalid request body or captcha not found
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "401":
          description: Invalid captcha input
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      summary: Verify captcha input
      tags:
      - captcha
...
  /wallets:
    get:
      description: دریافت لیست کیف پول‌های یک کاربر
      parameters:
      - description: شناسه کاربر
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: لیست کیف پول‌ها
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: خطای داخلی سرور
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: لیست کیف پول‌ها
      tags:
      - wallet
  /wallets/{wallet_id}/withdraw:
    post:
      consumes:
      - application/json
      description: برداشت مبلغ مشخص از کیف پول یک کاربر
      parameters:
      - description: شناسه کیف پول
        in: path
        name: wallet_id
        required: true
        type: string
      - description: اطلاعات برداشت
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exchange_internal_model.WalletWithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: برداشت موفق
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "403":
          description: دسترسی غیرمجاز
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "404":
          description: کیف پول پیدا نشد
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: خطای داخلی سرور
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: برداشت از کیف پول
      tags:
      - wallet
  /wallets/deposit:
    post:
      consumes:
      - application/json
      description: واریز مبلغ مشخص به کیف پول یک کاربر
      parameters:
      - description: شناسه کیف پول
        in: path
        name: wallet_id
        required: true
        type: string
      - description: اطلاعات واریز
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exchange_internal_model.WalletDepositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: واریز موفق
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "403":
          description: دسترسی غیرمجاز
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "404":
          description: کیف پول پیدا نشد
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: خطای داخلی سرور
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: واریز به کیف پول
      tags:
      - wallet
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"